requirements.txt
Django==1.5.2
MySQL-python==1.2.3
djangorestframework==2.3.8
yolk==0.4.3
